backend/tests/test_basic_agentic_functionality.py
"""
Basic Agentic Functionality Tests
Tests the core autonomous capabilities without requiring full dependencies
"""
import pytest
from app.core.data_iteration_engine import DataIterationEngine
from app.core.agent_self_iteration import AgentSelfIterationEngine
from app.core.agent_communication import AgentCommunicationRouter
from app.core.health_monitor import HealthMonitor

class TestBasicAgenticFunctionality:
    """Test basic agentic functionality"""
    
    def test_data_iteration_engine_instantiation(self):
        """Test that DataIterationEngine can be instantiated"""
        engine = DataIterationEngine()
        assert engine is not None
        assert hasattr(engine, 'analyze_data_quality')
        assert hasattr(engine, 'optimize_data_pipeline')
        
    def test_agent_self_iteration_engine_instantiation(self):
        """Test that AgentSelfIterationEngine can be instantiated"""
        engine = AgentSelfIterationEngine()
        assert engine is not None
        assert hasattr(engine, 'evaluate_agent_performance')
        assert hasattr(engine, 'evolve_agent_strategy')
        
    def test_agent_communication_router_instantiation(self):
        """Test that AgentCommunicationRouter can be instantiated"""
        router = AgentCommunicationRouter()
        assert router is not None
        assert hasattr(router, 'route_message')
        assert hasattr(router, 'build_consensus')
        
    def test_health_monitor_instantiation(self):
        """Test that HealthMonitor can be instantiated"""
        monitor = HealthMonitor()
        assert monitor is not None
        assert hasattr(monitor, 'check_system_health')
        assert hasattr(monitor, 'detect_anomalies')
        
    def test_data_quality_analysis(self):
        """Test basic data quality analysis"""
        engine = DataIterationEngine()
        test_data = {"field1": "value1", "field2": "value2"}
        result = engine._calculate_completeness(test_data)
        assert result == 1.0
        
    def test_empty_data_quality_analysis(self):
        """Test data quality analysis with empty data"""
        engine = DataIterationEngine()
        test_data = {}
        result = engine._calculate_completeness(test_data)
        assert result == 0.0