backend/requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
slowapi==0.1.9
PyJWT==2.8.0
gunicorn==23.0.0
pydantic==2.9.2
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.2.0
python-multipart==0.0.9
sqlalchemy==2.0.36
psycopg2-binary==2.9.9
redis==5.2.0
chromadb==0.4.24
langchain==0.2.12
langchain-core==0.2.30
langchain-community==0.2.12
langchain-openai==0.1.21
langchain-anthropic==0.1.15
langchain-google-genai==1.0.8
sentence-transformers==3.0.1
numpy==1.26.4
pandas==2.2.2
asyncio==3.4.3
httpx==0.27.0
celery==5.4.0
flower==2.0.1